working with network call

+++ app/src/main/java/com/assa/barcodereader/entity/Product.kt
@@ -3,6 +3,7 @@ package com.assa.barcodereader.entity
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import java.math.BigDecimal
 import java.time.LocalDate
 
 @Entity
@@ -13,6 +14,6 @@ data class Product(
     @ColumnInfo(name = "box_number") var boxNumber: String,
     @ColumnInfo(name = "amount") var amount: String,
     @ColumnInfo(name = "best_before") var bestBefore: String,
-    @ColumnInfo(name = "weight") var weight: Double,
+    @ColumnInfo(name = "weight") var weight: BigDecimal,
     @ColumnInfo(name = "transaction_number") var transactionNumber: Double
 )
